docs: fix styling of code blocks
After adding the `pymdownx.superfences` extension in d904bcd, styling of
code blocks is being handled by `pymdownx.highlight` instead of
`codehilite`. This means that our custom CSS rules are not working
anymore. This commit removes the superseded `codehilite` extension and
fixes the CSS rules to work with `pymdownx.highlight`.